[#160] Changed currency from double to long in module Tax

+++ tax/src/main/java/pt/ulisboa/tecnico/softeng/tax/domain/Invoice.java
@@ -7,7 +7,8 @@ import pt.ulisboa.tecnico.softeng.tax.exception.TaxException;
 
 public class Invoice extends Invoice_Base {
 
-	public Invoice(double value, LocalDate date, ItemType itemType, TaxPayer seller, TaxPayer buyer, DateTime time) {
+	public Invoice(long value, LocalDate date, ItemType itemType, TaxPayer seller, TaxPayer buyer, DateTime time) {
+  
 		checkArguments(value, date, itemType, seller, buyer, time);
 
 		setReference(Integer.toString(seller.getIrs().getCounter()));
@@ -19,12 +20,13 @@ public class Invoice extends Invoice_Base {
 		setBuyer(buyer);
 		setTime(time);
 
-		setIva(value * itemType.getTax() / 100);
+		setIva(value * itemType.getTax() / 100.0);
 
 		setIrs(getSeller().getIrs());
 	}
 
-	public Invoice(double value, LocalDate date, ItemType itemType, TaxPayer seller, TaxPayer buyer) {
+	public Invoice(long value, LocalDate date, ItemType itemType, TaxPayer seller, TaxPayer buyer) {
+
 		this(value, date, itemType, seller, buyer, DateTime.now());
 	}
 
@@ -36,9 +38,9 @@ public class Invoice extends Invoice_Base {
 		deleteDomainObject();
 	}
 
-	private void checkArguments(double value, LocalDate date, ItemType itemType, TaxPayer seller, TaxPayer buyer,
+	private void checkArguments(long value, LocalDate date, ItemType itemType, TaxPayer seller, TaxPayer buyer,
 			DateTime time) {
-		if (value <= 0.0f) {
+		if (value <= 0) {
 			throw new TaxException();
 		}
